+++ app/Models/Feedback.php
@@ -5,10 +5,11 @@ namespace App\Models;
 use App\Http\Traits\HasUuid;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Feedback extends Model
 {
-    use HasFactory, HasUuid;
+    use HasFactory, HasUuid, SoftDeletes;
 
     public const CATEGORIES = ['bug', 'feature_request', 'general'];